refactor for reuse in forms
