+++ modules/vba_documents/app/workers/vba_documents/deployments_notifier.rb
@@ -13,7 +13,7 @@ module VBADocuments
       result = nil
       begin
         VBADocuments::GitItems.populate
-        result = VBADocuments::GitItems.notify
+        result = VBADocuments::GitItems.notify(Settings.vba_documents.slack.notification_url)
       rescue => e
         Rails.logger.error('Failed to notify of new VBA document deployments', e)
         result = e
